requirements.txt
Flask
twilio
vosk
openai
requests
langdetect
python-dotenv
pydub